+++ spring-boot-data-jpa-actual/src/main/java/com/fayzal/springboot/datajpaactual/app/models/dao/IClienteDao.java
@@ -10,7 +10,7 @@ public interface IClienteDao {
 	
 	public void save(Cliente cliente);
 	
-	public Cliente findOne(Integer id);
+	public Cliente findOne(Long id);
 	
-	public void delete(Integer id);
+	public void delete(Long id);
 }
