+++ spring-boot-data-jpa-actual/src/main/java/com/fayzal/springboot/datajpaactual/app/service/IClienteService.java
@@ -15,8 +15,8 @@ public interface IClienteService {
 	
 	public void save(Cliente cliente);
 	
-	public Cliente findOne(Integer id);
+	public Cliente findOne(Long id);
 	
-	public void delete(Integer id);
+	public void delete(Long id);
 	
 }
